Feat: EditPartyDto, Move eat_party into party
