+++ src/party/party.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { EatParty } from 'src/entity/eat_party.entity';
+import { EatParty } from 'src/party/eat_party.entity';
 import { User } from 'src/entity/user.entity';
 import { PartyController } from './party.controller';
 import { PartyService } from './party.service';
